backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
motor==3.3.2
sentence-transformers>=2.3.0,<3.0.0
transformers>=4.34.0,<5.0.0
torch>=1.11.0
numpy>=1.24.3,<2.0.0
scikit-learn>=1.3.2
openai==1.3.7
Pillow>=10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
huggingface-hub>=0.20.0,<1.0.0

